#485 Fix order email total rendering
In Luma, the totals.phtml and the tax.phtml are used both for order
emails and for the customer order history.
In the past, Hyvä used the same template for both, too, which caused
either the frontend or the emails not to be rendered properly.

This PR moves the templates that are used in the store front to a new
location, which means the standard Luma templates are used for the order
emails.

This is a BC breaking change, but I am not aware of a cleaner way to fix this.

Closes #485