#485 Fix order email total rendering (1.1.x)
In Luma, the totals.phtml and the tax.phtml are used both for order
emails and for the customer order history.

In the past, Hyvä used the same template for both, too, which caused
either the frontend or the emails not to be rendered properly.

This PR moves the templates that are used in the store front to a new
location, which means the standard Luma templates are used for the order
emails.

This is the accompanying MR for the 1.2.x MR https://gitlab.hyva.io/hyva-themes/magento2-default-theme/-/merge_requests/582